requirements.txt
numpy
opencv_contrib_python
opencv_python
tensorflo
matplotlib
kaggle
seaborn
dlib